pyproject.toml
[tool.poetry]
name = "bashbot"
version = "0.1.0"
description = ""
authors = ["Commo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
python-telegram-bot = "^20.1"
openai = "^0.27.0"
pydantic = {extras = ["dotenv", "email"], version = "^1.10.5"}
fastapi = "^0.92.0"
django = "^4.1.7"
django-extensions = "^3.2.1"
requests = "^2.28.2"
psycopg2-binary = "^2.9.3"
django-cities-light = "^3.9.0"
django-phonenumber-field = "^6.0.0"
phonenumbers = "^8.12.43"
fastapi-jwt-auth = "^0.5.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
Pillow = "9.0.1"
python-multipart = "^0.0.5"
fastapi-cache2 = "^0.2.1"
django-debug-toolbar = "^3.8.1"
django-leaflet = "^0.28.3"
uvicorn = "^0.20.0"
redis-om = "^0.1.2"
geojson-pydantic = "^0.5.0"
itsdangerous = "^2.1.2"
whitenoise = "^6.4.0"
numpy = "^1.24.2"
redis = "^4.5.1"
fief-client = {extras = ["fastapi"], version = "^0.16.1"}


[tool.poetry.dev-dependencies]
# black = "^21.7b0"
# django-stubs = {git = "https://github.com/typeddjango/django-stubs.git"}
# flake8 = "^3.9.2"
# mypy = "^0.910"
pytest = "^6.2.2"
# pandas = "^1.2.4"
#jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

